Refactored ManagementApiComposer into several extension classes
